Add a registry_client module and factor out registry_endpoint
