+++ utils/registry_client.py
@@ -1,0 +1,14 @@
+import requests
+
+def registry_endpoint(url):
+    page = url
+    while page:
+        response = requests.get(url=page)
+        response.raise_for_status()
+        page = response.json().get('meta', {}).get('next', None)
+        if page:
+            page = f"https://registry.cdlib.org{page}"
+
+        collections = response.json().get('objects', [response.json()])
+        for collection in collections:
+            yield collection
